refactor: delegate json object behavior
